#OSIS-869 Refactoring
