+++ base/forms/learning_units.py
@@ -26,15 +26,14 @@
 
 from django import forms
 from django.core.exceptions import ValidationError
-from django.db.models import Prefetch
 from django.utils.translation import ugettext_lazy as _
 
 from base import models as mdl
-from base.business.entity import get_entities_ids, get_entity_container_list
+from base.business.entity import get_entities_ids, get_entity_container_list, build_entity_container_prefetch
 from base.business.entity_version import SERVICE_COURSE
 from base.business.learning_unit_year_with_context import append_latest_entities
 from base.forms.common import get_clean_data, treat_empty_or_str_none_as_none, TooManyResultsException
-from base.models import entity_version as mdl_entity_version, learning_unit_year
+from base.models import learning_unit_year
 from base.models.enums import entity_container_year_link_type, learning_container_year_types, \
     learning_unit_year_subtypes, active_status
 from base.forms.learning_unit_search import SearchForm
@@ -166,20 +165,3 @@ def get_filter_learning_container_ids(filter_data):
                                                      entity_container_year_link_type.ALLOCATION_ENTITY)
 
     return entities_id_list if entities_id_list else None
-
-
-def build_entity_container_prefetch():
-    parent_version_prefetch = Prefetch('parent__entityversion_set',
-                                       queryset=mdl_entity_version.search(),
-                                       to_attr='entity_versions')
-    entity_version_prefetch = Prefetch('entity__entityversion_set',
-                                       queryset=mdl_entity_version.search()
-                                       .prefetch_related(parent_version_prefetch),
-                                       to_attr='entity_versions')
-    entity_container_prefetch = Prefetch('learning_container_year__entitycontaineryear_set',
-                                         queryset=mdl.entity_container_year.search(
-                                             link_type=[entity_container_year_link_type.ALLOCATION_ENTITY,
-                                                        entity_container_year_link_type.REQUIREMENT_ENTITY])
-                                         .prefetch_related(entity_version_prefetch),
-                                         to_attr='entity_containers_year')
-    return entity_container_prefetch
