+++ base/views/learning_units/search.py
@@ -30,7 +30,8 @@ from django.forms import formset_factory
 from django.shortcuts import get_object_or_404
 from django.utils.translation import ugettext_lazy as _
 
-from base.business.learning_unit import SERVICE_COURSES_SEARCH, create_xls, get_last_academic_years, SIMPLE_SEARCH
+from base.business.learning_unit import SERVICE_COURSES_SEARCH, create_xls, get_last_academic_years, SIMPLE_SEARCH, \
+    get_learning_units_summary_status
 from base.forms.common import TooManyResultsException
 from base.forms.learning_unit_create import MAX_RECORDS
 from base.forms.learning_units import LearningUnitYearForm
@@ -41,16 +42,6 @@ from base.models.person import Person, find_by_user
 from base.views import layout
 from base.views.common import check_if_display_message, display_error_messages, display_success_messages
 from base.business import learning_unit_proposal as proposal_business
-from base.models import learning_unit_year
-from attribution.models import attribution_new
-from cms.models import translated_text
-from cms.enums.entity_name import LEARNING_UNIT_YEAR
-from base.business.learning_unit import CMS_LABEL_PEDAGOGY
-from base.business.entity import get_entities_ids, get_entity_container_list, get_entity_calendar
-from base.models import entity_version as mdl_entity_version
-from base import models as mdl
-from base.models.enums import entity_container_year_link_type
-from django.db.models import Prefetch
 
 PROPOSAL_SEARCH = 3
 SUMMARY_LIST = 4
@@ -177,45 +168,10 @@ def _force_state(formset, request):
         display_error_messages(request, _("error_modification_learning_unit"))
 
 
-def build_entity_container_prefetch():
-    parent_version_prefetch = Prefetch('parent__entityversion_set',
-                                       queryset=mdl_entity_version.search(),
-                                       to_attr='entity_versions')
-    entity_version_prefetch = Prefetch('entity__entityversion_set',
-                                       queryset=mdl_entity_version.search()
-                                       .prefetch_related(parent_version_prefetch),
-                                       to_attr='entity_versions')
-    entity_container_prefetch = Prefetch('learning_container_year__entitycontaineryear_set',
-                                         queryset=mdl.entity_container_year.search(
-                                             link_type=[entity_container_year_link_type.ALLOCATION_ENTITY,
-                                                        entity_container_year_link_type.REQUIREMENT_ENTITY])
-                                         .prefetch_related(entity_version_prefetch),
-                                         to_attr='entity_containers_year')
-    return entity_container_prefetch
-
-
 @login_required
 @permission_required('base.can_access_learningunit', raise_exception=True)
 def learning_units_summary_list(request):
-    a_person = find_by_user(request.user)
-    entities_version_attached = a_person.find_main_entities_version
-    learning_units_found = []
-    current_academic_yr = current_academic_year()
-    for an_entity_version in entities_version_attached:
-        a_calendar = get_entity_calendar(an_entity_version, current_academic_yr)
-        if a_calendar:
-            entity_learning_unit_yr_list = get_list_entity_learning_unit_yr(an_entity_version, current_academic_yr)
-            if entity_learning_unit_yr_list:
-                for lu in entity_learning_unit_yr_list:
-                    lu.summary_responsibles = attribution_new.search(summary_responsible=True,
-                                                                     learning_container_year=lu.learning_container_year)
-                    lu.summary_status = translated_text.check_changed(LEARNING_UNIT_YEAR,
-                                                                      a_calendar.start_date,
-                                                                      a_calendar.end_date,
-                                                                      lu.id,
-                                                                      CMS_LABEL_PEDAGOGY)
-                learning_units_found.extend(entity_learning_unit_yr_list)
-
+    learning_units_found = get_learning_units_summary_status(find_by_user(request.user))
     context = {
         'learning_units': sorted(learning_units_found, key=lambda t: t.acronym),
         'experimental_phase': True,
@@ -223,16 +179,3 @@ def learning_units_summary_list(request):
     }
 
     return layout.render(request, "learning_units.html", context)
-
-
-def get_list_entity_learning_unit_yr(an_entity_version, current_academic_yr):
-    entity_ids = get_entities_ids(an_entity_version.entity.most_recent_acronym, True)
-    entities_id_list = get_entity_container_list([], entity_ids, entity_container_year_link_type.REQUIREMENT_ENTITY)
-
-    return learning_unit_year.search(**{'learning_container_year_id': entities_id_list,
-                                        'academic_year_id': current_academic_yr,
-                                        'status': True}) \
-        .select_related('academic_year', 'learning_container_year',
-                        'learning_container_year__academic_year') \
-        .prefetch_related(build_entity_container_prefetch()) \
-        .order_by('academic_year__year', 'acronym')
